+++ components/motor/esp_simplefoc/port/esp/esp_app_fmath.cpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2023-2024 Espressif Systems (Shanghai) CO LTD
+ * SPDX-FileCopyrightText: 2023 Espressif Systems (Shanghai) CO LTD
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -9,14 +9,12 @@
 #include "soc/soc_caps.h"
 #include "IQmathLib.h"
 
-
 float _sin(float a)
 {
     _iq iq_radians = _IQ(a);
     _iq result = _IQsin(iq_radians);
     return _IQtoF(result);
 }
-
 
 float _cos(float a)
 {
@@ -25,19 +23,16 @@ float _cos(float a)
     return _IQtoF(result);
 }
 
-
 float _normalizeAngle(float angle)
 {
     float a = fmod(angle, _2PI);
     return a >= 0 ? a : (a + _2PI);
 }
 
-
 float _electricalAngle(float shaft_angle, int pole_pairs)
 {
     return (shaft_angle * pole_pairs);
 }
-
 
 float _sqrtApprox(float number)
 {