+++ components/motor/esp_simplefoc/port/esp/include/esp_simplefoc.h
@@ -13,4 +13,5 @@
 #include "sensors/GenericSensor.h"
 #include "current_sense/GenericCurrentSense.h"
 #include "current_sense/LowsideCurrentSense.h"
-#include "../../angle_sensor/sensor_as5600.h"
+#include "../../angle_sensor/as5600.h"
+#include "../../angle_sensor/mt6701.h"